fix: 重新组织了服务代码，并修复的事务问题和用户信息服务的逻辑问题

+++ src/service/user.service.ts
@@ -1,8 +1,6 @@
 import { PrismaClient, User, UserRole } from '@prisma/client'
-import * as uuid from 'uuid'
 import { ResponseError } from '../util/errors'
 import sharp from 'sharp'
-import fs from 'fs'
 import OSSService from './oss.service'
 import { asyncInitializeRoutine } from '../app/container'
 import { Logger } from 'pino'
@@ -27,7 +25,6 @@ export default class UserService {
                 this.logger.info('Checking OSS avatar bucket')
                 if (await this.ossService.createBucketIfNotExist('avatar')) {
                     this.logger.info('Avatar bucket does not exist, creating...')
-                    await this.uploadUserAvatar('default', fs.readFileSync('assets/default-avatar.svg'))
                 }
                 this.logger.info('OSS avatar bucket OK')
             } catch (err) {
@@ -37,33 +34,13 @@ export default class UserService {
         })
     }
 
-    hasModifyPermissionOrElse403(currentUser: User, userId: string) {
-        if (currentUser.role === UserRole.ADMIN) {
-            return
-        }
-        if (!uuid.validate(userId) || !uuid.parse(userId).every(v => v === 0)) {
-            throw new ResponseError(403, 'Permission denied')
-        }
+    async getUser(id: string) {
+        return await this.prisma.user.findUnique({ where: { id } })
     }
 
-    getTargetId(currentUser: User, userId: string) {
-        return uuid.validate(userId) && uuid.parse(userId).every(v => v === 0) ? currentUser.id : userId
-    }
-
-    hasRoleModifyPermissionOrElse403(currentUser: User, role?: UpdateUserProfile['role']) {
-        if (currentUser.role === UserRole.ADMIN) {
-            return
-        }
-        if (role === 'admin') {
-            throw new ResponseError(403, 'Permission denied')
-        }
-    }
-
-    async updateUserProfile(id: string, name?: string, description?: string, role?: UpdateUserProfile['role'], emailVisible?: boolean, createdAtVisible?: boolean): Promise<User> {
+    async updateUserProfile(currentUserId: string, id: string, name?: string, description?: string, role?: UpdateUserProfile['role'], emailVisible?: boolean, createdAtVisible?: boolean): Promise<User> {
         let roleKey: UserRole | undefined
-        if (role === undefined) {
-            roleKey = undefined
-        } else if (role === 'customer') {
+        if (role === 'customer') {
             roleKey = UserRole.USER
         } else if (role === 'rider') {
             roleKey = UserRole.RIDER
@@ -72,27 +49,49 @@ export default class UserService {
         } else if (role === 'admin') {
             roleKey = UserRole.ADMIN
         } else {
-            throw new Error('Unreachable')
+            roleKey = undefined
         }
-        return await this.prisma.user.update({
-            where: { id },
-            data: { name, description, role: roleKey, emailVisible, createdAtVisible },
+        return await this.prisma.$transaction(async tx => {
+            const user = await tx.user.findUnique({ where: { id } })
+            if (!user) {
+                throw new ResponseError(404, 'User not found')
+            }
+            const currentUser = await tx.user.findUnique({ where: { id: currentUserId } })
+            if (!currentUser || (currentUser.role !== UserRole.ADMIN && currentUser.id !== id)) {
+                throw new ResponseError(403, 'Permission denied')
+            }
+            if (currentUser.role !== UserRole.ADMIN && role === 'admin') {
+                throw new ResponseError(403, 'Permission denied')
+            }
+            return await tx.user.update({
+                where: { id },
+                data: { name, description, role: roleKey, emailVisible, createdAtVisible },
+            })
         })
     }
 
     readonly ossContentType = 'image/webp'
 
-    async getUserAvatarLinks(id: string): Promise<{ origin: string, thumbnail: string }> {
-        if (await this.ossService.existsObject('avatar', `${id}.webp`)) {
-            const [origin, thumbnail] = await Promise.all([this.ossService.getObjectUrl('avatar', `${id}.webp`), this.ossService.getObjectUrl('avatar', `${id}-thumbnail.webp`)])
-            return { origin, thumbnail }
-        } else {
-            const [origin, thumbnail] = await Promise.all([this.ossService.getObjectUrl('avatar', `default.webp`), this.ossService.getObjectUrl('avatar', `default-thumbnail.webp`)])
-            return { origin, thumbnail }
-        }
+    private async checkModifyAvatarPermission(currentUserId: string, id: string) {
+        return await this.prisma.$transaction(async tx => {
+            const user = await tx.user.findUnique({ where: { id } })
+            if (!user) {
+                throw new ResponseError(404, 'User not found')
+            }
+            const currentUser = await tx.user.findUnique({ where: { id: currentUserId } })
+            if (!currentUser || (currentUser.role !== UserRole.ADMIN && currentUser.id !== id)) {
+                throw new ResponseError(403, 'Permission denied')
+            }
+        })
     }
 
-    async uploadUserAvatar(id: string, buffer: Buffer): Promise<{ origin: string, thumbnail: string }> {
+    async getUserAvatarLinks(id: string): Promise<{ origin: string, thumbnail: string }> {
+        const [origin, thumbnail] = await Promise.all([this.ossService.getObjectUrl('avatar', `${id}.webp`), this.ossService.getObjectUrl('avatar', `${id}-thumbnail.webp`)])
+        return { origin, thumbnail }
+    }
+
+    async uploadUserAvatar(currentUserId: string, id: string, buffer: Buffer): Promise<{ origin: string, thumbnail: string }> {
+        await this.checkModifyAvatarPermission(currentUserId, id)
         const [origin, thumbnail] = await Promise.all([
             this.ossService.putObject('avatar', `${id}.webp`, sharp(buffer).toFormat('webp'), this.ossContentType),
             this.ossService.putObject('avatar', `${id}-thumbnail.webp`, sharp(buffer).resize(128, 128).toFormat('webp'), this.ossContentType)
@@ -100,7 +99,8 @@ export default class UserService {
         return { origin, thumbnail }
     }
 
-    async deleteUserAvatar(id: string) {
+    async deleteUserAvatar(currentUserId: string, id: string) {
+        await this.checkModifyAvatarPermission(currentUserId, id)
         await Promise.all([
             this.ossService.removeObject('avatar', `${id}.webp`),
             this.ossService.removeObject('avatar', `${id}-thumbnail.webp`)
